@@ -26,8 +26,8 @@ import 'scroll_view.dart';
 class PositionedList extends StatefulWidget {
   /// Create a [PositionedList].
   const PositionedList({
-    @required this.itemCount,
-    @required this.itemBuilder,
+    required this.itemCount,
+    required this.itemBuilder,
     this.findChildIndexCallback,
     this.separatorBuilder,
     this.controller,
@@ -56,16 +56,16 @@ class PositionedList extends StatefulWidget {
 
   /// If not null, called to build separators for between each item in the list.
   /// Called with 0 <= index < itemCount - 1.
-  final IndexedWidgetBuilder separatorBuilder;
+  final IndexedWidgetBuilder? separatorBuilder;
 
-  final ChildIndexGetter findChildIndexCallback;
+  final ChildIndexGetter? findChildIndexCallback;
 
   /// An object that can be used to control the position to which this scroll
   /// view is scrolled.
-  final ScrollController controller;
+  final ScrollController? controller;
 
   /// Notifier that reports the items laid out in the list after each frame.
-  final ItemPositionsNotifier itemPositionsNotifier;
+  final ItemPositionsNotifier? itemPositionsNotifier;
 
   /// Index of an item to initially align to a position within the viewport
   /// defined by [alignment].
@@ -75,7 +75,7 @@ class PositionedList extends StatefulWidget {
   /// should be placed.
   ///
   /// See [ItemScrollController.jumpTo] for an explanation of alignment.
-  final double alignment;
+  final double? alignment;
 
   /// The axis along which the scroll view scrolls.
   ///
@@ -95,15 +95,15 @@ class PositionedList extends StatefulWidget {
   /// user stops dragging the scroll view.
   ///
   /// See [ScrollView.physics].
-  final ScrollPhysics physics;
+  final ScrollPhysics? physics;
 
   /// {@macro flutter.widgets.scrollable.cacheExtent}
-  final double cacheExtent;
+  final double? cacheExtent;
 
   /// The number of children that will contribute semantic information.
   ///
   /// See [ScrollView.semanticChildCount] for more information.
-  final int semanticChildCount;
+  final int? semanticChildCount;
 
   /// Whether to wrap each child in an [IndexedSemantics].
   ///
@@ -111,7 +111,7 @@ class PositionedList extends StatefulWidget {
   final bool addSemanticIndexes;
 
   /// The amount of space by which to inset the children.
-  final EdgeInsets padding;
+  final EdgeInsets? padding;
 
   /// Whether to wrap each child in a [RepaintBoundary].
   ///
@@ -130,8 +130,8 @@ class PositionedList extends StatefulWidget {
 class _PositionedListState extends State<PositionedList> {
   final Key _centerKey = UniqueKey();
 
-  final registeredElements = ValueNotifier<Set<Element>>(null);
-  ScrollController scrollController;
+  final registeredElements = ValueNotifier<Set<Element>?>(null);
+  ScrollController? scrollController;
 
   bool updateScheduled = false;
 
@@ -139,13 +139,13 @@ class _PositionedListState extends State<PositionedList> {
   void initState() {
     super.initState();
     scrollController = widget.controller ?? ScrollController();
-    scrollController.addListener(_schedulePositionNotificationUpdate);
+    scrollController!.addListener(_schedulePositionNotificationUpdate);
     _schedulePositionNotificationUpdate();
   }
 
   @override
   void dispose() {
-    scrollController.removeListener(_schedulePositionNotificationUpdate);
+    scrollController!.removeListener(_schedulePositionNotificationUpdate);
     super.dispose();
   }
 
@@ -180,9 +180,9 @@ class _PositionedListState extends State<PositionedList> {
                       : _buildSeparatedListElement(2 * widget.positionedIndex - (index + 1)),
                   childCount: widget.separatorBuilder == null ? widget.positionedIndex : 2 * widget.positionedIndex,
                   findChildIndexCallback: (key) {
-                    int rawIndex;
+                    int? rawIndex;
                     if (widget.findChildIndexCallback != null) {
-                      final int index = widget.findChildIndexCallback(key);
+                      final int? index = widget.findChildIndexCallback!(key);
                       if (index != null) {
                         if (key.toString().contains('root')) {
                           rawIndex = -((index * 2) - (2 * widget.positionedIndex)) - 1;
@@ -209,9 +209,9 @@ class _PositionedListState extends State<PositionedList> {
                     : _buildSeparatedListElement(index + 2 * widget.positionedIndex),
                 childCount: widget.itemCount != 0 ? 1 : 0,
                 findChildIndexCallback: (key) {
-                  int rawIndex;
+                  int? rawIndex;
                   if (widget.findChildIndexCallback != null) {
-                    final int index = widget.findChildIndexCallback(key);
+                    final int? index = widget.findChildIndexCallback!(key);
                     if (index != null) {
                       rawIndex = (index * 2) - (2 * widget.positionedIndex);
                     }
@@ -237,9 +237,9 @@ class _PositionedListState extends State<PositionedList> {
                       ? widget.itemCount - widget.positionedIndex - 1
                       : 2 * (widget.itemCount - widget.positionedIndex - 1),
                   findChildIndexCallback: (key) {
-                    int rawIndex;
+                    int? rawIndex;
                     if (widget.findChildIndexCallback != null) {
-                      final int index = widget.findChildIndexCallback(key);
+                      final int? index = widget.findChildIndexCallback!(key);
                       if (index != null) {
                         if (key.toString().contains('root')) {
                           rawIndex = (index * 2) - (2 * widget.positionedIndex) - 1;
@@ -265,7 +265,7 @@ class _PositionedListState extends State<PositionedList> {
     if (index.isEven) {
       return _buildItem(index ~/ 2);
     } else {
-      return widget.separatorBuilder(context, index ~/ 2);
+      return widget.separatorBuilder!(context, index ~/ 2);
     }
   }
 
@@ -294,21 +294,21 @@ class _PositionedListState extends State<PositionedList> {
   EdgeInsets get _centerSliverPadding => widget.scrollDirection == Axis.vertical
       ? widget.reverse
           ? widget.padding?.copyWith(
-                  top: widget.positionedIndex == widget.itemCount - 1 ? widget.padding.top : 0,
-                  bottom: widget.positionedIndex == 0 ? widget.padding.bottom : 0) ??
+                  top: widget.positionedIndex == widget.itemCount - 1 ? widget.padding!.top : 0,
+                  bottom: widget.positionedIndex == 0 ? widget.padding!.bottom : 0) ??
               EdgeInsets.all(0)
           : widget.padding?.copyWith(
-                  top: widget.positionedIndex == 0 ? widget.padding.top : 0,
-                  bottom: widget.positionedIndex == widget.itemCount - 1 ? widget.padding.bottom : 0) ??
+                  top: widget.positionedIndex == 0 ? widget.padding!.top : 0,
+                  bottom: widget.positionedIndex == widget.itemCount - 1 ? widget.padding!.bottom : 0) ??
               EdgeInsets.all(0)
       : widget.reverse
           ? widget.padding?.copyWith(
-                  left: widget.positionedIndex == widget.itemCount - 1 ? widget.padding.left : 0,
-                  right: widget.positionedIndex == 0 ? widget.padding.right : 0) ??
+                  left: widget.positionedIndex == widget.itemCount - 1 ? widget.padding!.left : 0,
+                  right: widget.positionedIndex == 0 ? widget.padding!.right : 0) ??
               EdgeInsets.all(0)
           : widget.padding?.copyWith(
-                left: widget.positionedIndex == 0 ? widget.padding.left : 0,
-                right: widget.positionedIndex == widget.itemCount - 1 ? widget.padding.right : 0,
+                left: widget.positionedIndex == 0 ? widget.padding!.left : 0,
+                right: widget.positionedIndex == widget.itemCount - 1 ? widget.padding!.right : 0,
               ) ??
               EdgeInsets.all(0);
 
@@ -323,52 +323,48 @@ class _PositionedListState extends State<PositionedList> {
   void _schedulePositionNotificationUpdate() {
     if (!updateScheduled) {
       updateScheduled = true;
-      SchedulerBinding.instance.addPostFrameCallback((_) {
+      SchedulerBinding.instance!.addPostFrameCallback((_) {
         if (registeredElements.value == null) {
           updateScheduled = false;
           return;
         }
         final positions = <ItemPosition>[];
-        RenderViewport viewport;
-        for (var element in registeredElements.value) {
+        RenderViewport? viewport;
+        for (var element in registeredElements.value!) {
           try {
-            final RenderBox box = element.renderObject;
+            final RenderBox? box = element.renderObject as RenderBox?;
             if (box == null) continue;
-            viewport ??= RenderAbstractViewport.of(box);
+            viewport ??= RenderAbstractViewport.of(box) as RenderViewport?;
             if (viewport == null) continue;
-            final ValueKey<String> key = element.widget.key;
+            final ValueKey<String>? key = element.widget.key as ValueKey<String>?;
             int index = 0;
             if (widget.findChildIndexCallback != null) {
-              index = widget.findChildIndexCallback(key) ?? 0;
+              index = widget.findChildIndexCallback!(key!) ?? 0;
             }
             if (widget.scrollDirection == Axis.vertical) {
-              final reveal = viewport
-                  .getOffsetToReveal(box, 0)
-                  .offset;
+              final reveal = viewport.getOffsetToReveal(box, 0).offset;
               final itemOffset = reveal - viewport.offset.pixels + viewport.anchor * viewport.size.height;
               positions.add(ItemPosition(
-                // index: key.value,
+                  // index: key.value,
                   index: index,
-                  itemLeadingEdge: itemOffset.round() / scrollController.position.viewportDimension,
+                  itemLeadingEdge: itemOffset.round() / scrollController!.position.viewportDimension,
                   itemTrailingEdge:
-                  (itemOffset + box.size.height).round() / scrollController.position.viewportDimension));
+                      (itemOffset + box.size.height).round() / scrollController!.position.viewportDimension));
             } else {
-              final itemOffset = box
-                  .localToGlobal(Offset.zero, ancestor: viewport)
-                  .dx;
+              final itemOffset = box.localToGlobal(Offset.zero, ancestor: viewport).dx;
               positions.add(ItemPosition(
-                // index: key.value,
+                  // index: key.value,
                   index: index,
                   itemLeadingEdge: (widget.reverse
-                      ? scrollController.position.viewportDimension - (itemOffset + box.size.width)
-                      : itemOffset)
-                      .round() /
-                      scrollController.position.viewportDimension,
+                              ? scrollController!.position.viewportDimension - (itemOffset + box.size.width)
+                              : itemOffset)
+                          .round() /
+                      scrollController!.position.viewportDimension,
                   itemTrailingEdge: (widget.reverse
-                      ? scrollController.position.viewportDimension - itemOffset
-                      : (itemOffset + box.size.width))
-                      .round() /
-                      scrollController.position.viewportDimension));
+                              ? scrollController!.position.viewportDimension - itemOffset
+                              : (itemOffset + box.size.width))
+                          .round() /
+                      scrollController!.position.viewportDimension));
             }
           } catch (e) {
             print('[ChatListView] PositionedList _schedulePositionNotificationUpdate with error: $e');
