@@ -7,14 +7,14 @@ import 'package:flutter/widgets.dart';
 /// Widget whose [Element] calls a callback when the element is mounted.
 class PostMountCallback extends StatelessWidget {
   /// Creates a [PostMountCallback] widget.
-  const PostMountCallback({@required this.child, this.callback, Key key})
+  const PostMountCallback({required this.child, this.callback, Key? key})
       : super(key: key);
 
   /// The widget below this widget in the tree.
   final Widget child;
 
   /// Callback to call when the element for this widget is mounted.
-  final void Function() callback;
+  final void Function()? callback;
 
   @override
   StatelessElement createElement() => _PostMountCallbackElement(this);
@@ -27,9 +27,9 @@ class _PostMountCallbackElement extends StatelessElement {
   _PostMountCallbackElement(PostMountCallback widget) : super(widget);
 
   @override
-  void mount(Element parent, dynamic newSlot) {
+  void mount(Element? parent, dynamic newSlot) {
     super.mount(parent, newSlot);
-    final PostMountCallback postMountCallback = widget;
+    final PostMountCallback postMountCallback = widget as PostMountCallback;
     postMountCallback.callback?.call();
   }
 }
