@@ -13,14 +13,14 @@ typedef LatestMessageIdBuilder = String Function();
 
 class ChatListView extends StatefulWidget {
   const ChatListView.builder({
-    Key key,
-    @required this.firstLoadedBuilder,
-    @required this.latestMessageIdBuilder,
-    @required this.messageIds,
-    @required this.itemCount,
-    @required this.itemBuilder,
-    @required this.itemKeyPrefix,
-    @required this.separatorKeyPrefix,
+    Key? key,
+    required this.firstLoadedBuilder,
+    required this.latestMessageIdBuilder,
+    required this.messageIds,
+    required this.itemCount,
+    required this.itemBuilder,
+    required this.itemKeyPrefix,
+    required this.separatorKeyPrefix,
     this.listViewKey,
     this.initialScrollIndex = 0,
     this.initialAlignment = 0.0,
@@ -45,15 +45,15 @@ class ChatListView extends StatefulWidget {
         super(key: key);
 
   ChatListView.separated({
-    Key key,
-    @required this.firstLoadedBuilder,
-    @required this.latestMessageIdBuilder,
-    @required this.messageIds,
-    @required this.itemCount,
-    @required this.itemBuilder,
-    @required this.separatorBuilder,
-    @required this.itemKeyPrefix,
-    @required this.separatorKeyPrefix,
+    Key? key,
+    required this.firstLoadedBuilder,
+    required this.latestMessageIdBuilder,
+    required this.messageIds,
+    required this.itemCount,
+    required this.itemBuilder,
+    required this.separatorBuilder,
+    required this.itemKeyPrefix,
+    required this.separatorKeyPrefix,
     this.listViewKey,
     this.initialScrollIndex = 0,
     this.initialAlignment = 0.0,
@@ -77,7 +77,7 @@ class ChatListView extends StatefulWidget {
         assert(separatorBuilder != null),
         super(key: key);
 
-  final Key listViewKey;
+  final Key? listViewKey;
 
   final String itemKeyPrefix;
 
@@ -99,13 +99,13 @@ class ChatListView extends StatefulWidget {
 
   /// Called to build separators for between each item in the list.
   /// Called with 0 <= index < itemCount - 1.
-  final IndexedWidgetBuilder separatorBuilder;
+  final IndexedWidgetBuilder? separatorBuilder;
 
   /// Controller for jumping or scrolling to an item.
-  final ItemScrollController itemScrollController;
+  final ItemScrollController? itemScrollController;
 
   /// Notifier that reports the items laid out in the list after each frame.
-  final ItemPositionsListener itemPositionsListener;
+  final ItemPositionsListener? itemPositionsListener;
 
   /// Index of an item to initially align within the viewport.
   final int initialScrollIndex;
@@ -122,15 +122,15 @@ class ChatListView extends StatefulWidget {
   /// user stops dragging the scroll view.
   ///
   /// See [ScrollView.physics].
-  final ScrollPhysics physics;
+  final ScrollPhysics? physics;
 
   /// The number of children that will contribute semantic information.
   ///
   /// See [ScrollView.semanticChildCount] for more information.
-  final int semanticChildCount;
+  final int? semanticChildCount;
 
   /// The amount of space by which to inset the children.
-  final EdgeInsets padding;
+  final EdgeInsets? padding;
 
   /// Whether to wrap each child in an [IndexedSemantics].
   ///
@@ -154,28 +154,28 @@ class ChatListView extends StatefulWidget {
   /// scrolls, so using the [ScrollController.scrollTo] method may result
   /// in builds of widgets that would otherwise already be built in the
   /// cache extent.
-  final double minCacheExtent;
+  final double? minCacheExtent;
 
   /// Called when the [child] reaches the start of the list
-  final AsyncCallback onStartOfPage;
+  final AsyncCallback? onStartOfPage;
 
   /// Called when the [child] reaches the end of the list
-  final AsyncCallback onEndOfPage;
+  final AsyncCallback? onEndOfPage;
 
   /// Called when the list scrolling starts
-  final VoidCallback onPageScrollStart;
+  final VoidCallback? onPageScrollStart;
 
   /// Called when the list scrolling ends
-  final VoidCallback onPageScrollEnd;
+  final VoidCallback? onPageScrollEnd;
 
-  final OnPageAtBottom onPageAtBottom;
+  final OnPageAtBottom? onPageAtBottom;
 
   @override
   _ChatListViewState createState() => _ChatListViewState();
 }
 
 class _ChatListViewState extends State<ChatListView> {
-  LazyLoadScrollController _lazyLoadController;
+  LazyLoadScrollController? _lazyLoadController;
 
   @override
   void initState() {
@@ -201,7 +201,7 @@ class _ChatListViewState extends State<ChatListView> {
         },
         itemCount: widget.itemCount,
         itemBuilder: widget.itemBuilder,
-        separatorBuilder: widget.separatorBuilder,
+        separatorBuilder: widget.separatorBuilder!,
         findChildIndexCallback: (key) {
           if (key is ValueKey) {
             final String parsedKey =
